Split out views/reports.py
