+++ corehq/apps/hqadmin/views/reports.py
@@ -1,0 +1,264 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import unicode_literals
+
+import itertools
+import six.moves.html_parser
+import json
+import socket
+import uuid
+from io import StringIO
+from collections import defaultdict, namedtuple, OrderedDict, Counter
+from datetime import timedelta, date, datetime
+
+import dateutil
+from couchdbkit import ResourceNotFound
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth import login
+from django.contrib.auth.models import User
+from django.core import management, cache
+from django.core.exceptions import ObjectDoesNotExist
+from django.core.mail import mail_admins
+from django.http import (
+    HttpResponseRedirect,
+    HttpResponse,
+    HttpResponseBadRequest,
+    HttpResponseNotFound,
+    JsonResponse,
+    StreamingHttpResponse,
+)
+from django.http.response import Http404
+from django.shortcuts import render, redirect
+from django.template.loader import render_to_string
+from django.utils.decorators import method_decorator
+from django.utils.functional import cached_property
+from django.utils.translation import ugettext as _, ugettext_lazy
+from django.views.decorators.http import require_POST
+from django.views.generic import FormView, TemplateView, View
+from lxml import etree
+from lxml.builder import E
+from restkit import Resource
+from restkit.errors import Unauthorized
+
+from casexml.apps.case.models import CommCareCase
+from casexml.apps.phone.xml import SYNC_XMLNS
+from casexml.apps.stock.const import COMMTRACK_REPORT_XMLNS
+from corehq.apps.app_manager.models import ApplicationBase
+from corehq.apps.callcenter.indicator_sets import CallCenterIndicators
+from corehq.apps.callcenter.utils import CallCenterCase
+from corehq.apps.data_analytics.admin import MALTRowAdmin
+from corehq.apps.data_analytics.const import GIR_FIELDS
+from corehq.apps.data_analytics.models import MALTRow, GIRRow
+from corehq.apps.domain.auth import basicauth
+from corehq.apps.domain.decorators import (
+    require_superuser, require_superuser_or_contractor,
+    login_or_basic, domain_admin_required,
+    check_lockout)
+from corehq.apps.domain.models import Domain
+from corehq.apps.es import filters
+from corehq.apps.es.domains import DomainES
+from corehq.apps.hqadmin.reporting.exceptions import HistoTypeNotFoundException
+from corehq.apps.hqadmin.service_checks import run_checks
+from corehq.apps.hqwebapp.views import BaseSectionPageView
+from corehq.apps.locations.models import SQLLocation
+from corehq.apps.ota.views import get_restore_response, get_restore_params
+from corehq.apps.hqwebapp.decorators import use_datatables, use_jquery_ui, \
+    use_nvd3_v3
+from corehq.apps.users.models import CommCareUser, WebUser, CouchUser
+from corehq.apps.users.util import format_username
+from corehq.elastic import parse_args_for_es, run_query, ES_META
+from corehq.form_processor.backends.couch.dbaccessors import CaseAccessorCouch
+from corehq.form_processor.backends.sql.dbaccessors import CaseAccessorSQL
+from corehq.form_processor.exceptions import XFormNotFound, CaseNotFound
+from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
+from corehq.form_processor.models import XFormInstanceSQL, CommCareCaseSQL
+from corehq.form_processor.serializers import XFormInstanceSQLRawDocSerializer, \
+    CommCareCaseSQLRawDocSerializer
+from corehq.toggles import any_toggle_enabled, SUPPORT
+from corehq.util import reverse
+from corehq.util.couchdb_management import couch_config
+from corehq.util.supervisord.api import (
+    PillowtopSupervisorApi,
+    SupervisorException,
+    all_pillows_supervisor_status,
+    pillow_supervisor_status
+)
+from corehq.util.timer import TimingContext
+from couchforms.models import XFormInstance
+from couchforms.openrosa_response import RESPONSE_XMLNS
+from dimagi.utils.couch.database import get_db, is_bigcouch
+from dimagi.utils.csv import UnicodeWriter
+from dimagi.utils.dates import add_months
+from dimagi.utils.decorators.datespan import datespan_in_request
+from memoized import memoized
+from dimagi.utils.django.email import send_HTML_email
+from dimagi.utils.django.management import export_as_csv_action
+from dimagi.utils.parsing import json_format_date
+from dimagi.utils.web import json_response
+from corehq.apps.hqadmin.tasks import send_mass_emails
+from pillowtop.exceptions import PillowNotFoundError
+from pillowtop.utils import get_all_pillows_json, get_pillow_json, get_pillow_config_by_name
+from corehq.apps.hqadmin import service_checks, escheck
+from corehq.apps.hqadmin.forms import (
+    AuthenticateAsForm, BrokenBuildsForm, EmailForm, SuperuserManagementForm,
+    ReprocessMessagingCaseUpdatesForm,
+    DisableTwoFactorForm, DisableUserForm)
+from corehq.apps.hqadmin.history import get_recent_changes, download_changes
+from corehq.apps.hqadmin.models import HqDeploy
+from corehq.apps.hqadmin.reporting.reports import get_project_spaces, get_stats_data, HISTO_TYPE_TO_FUNC
+from corehq.apps.hqadmin.utils import get_celery_stats
+from corehq.apps.hqadmin.views.utils import BaseAdminSectionView, get_hqadmin_base_context
+import six
+from six.moves import filter
+
+
+@require_superuser
+@datespan_in_request(from_param="startdate", to_param="enddate", default_days=90)
+def stats_data(request):
+    histo_type = request.GET.get('histogram_type')
+    interval = request.GET.get("interval", "week")
+    datefield = request.GET.get("datefield")
+    get_request_params_json = request.GET.get("get_request_params", None)
+    get_request_params = (
+        json.loads(six.moves.html_parser.HTMLParser().unescape(get_request_params_json))
+        if get_request_params_json is not None else {}
+    )
+
+    stats_kwargs = {
+        k: get_request_params[k]
+        for k in get_request_params if k != "domain_params_es"
+    }
+    if datefield is not None:
+        stats_kwargs['datefield'] = datefield
+
+    domain_params_es = get_request_params.get("domain_params_es", {})
+
+    if not request.GET.get("enddate"):  # datespan should include up to the current day when unspecified
+        request.datespan.enddate += timedelta(days=1)
+
+    domain_params, __ = parse_args_for_es(request, prefix='es_')
+    domain_params.update(domain_params_es)
+
+    domains = get_project_spaces(facets=domain_params)
+
+    try:
+        return json_response(get_stats_data(
+            histo_type,
+            domains,
+            request.datespan,
+            interval,
+            **stats_kwargs
+        ))
+    except HistoTypeNotFoundException:
+        return HttpResponseBadRequest(
+            'histogram_type param must be one of <ul><li>{}</li></ul>'
+            .format('</li><li>'.join(HISTO_TYPE_TO_FUNC)))
+
+
+@require_superuser
+@datespan_in_request(from_param="startdate", to_param="enddate", default_days=365)
+def admin_reports_stats_data(request):
+    return stats_data(request)
+
+
+class DimagisphereView(TemplateView):
+
+    def get_context_data(self, **kwargs):
+        context = super(DimagisphereView, self).get_context_data(**kwargs)
+        context['tvmode'] = 'tvmode' in self.request.GET
+        return context
+
+
+def top_five_projects_by_country(request):
+    data = {}
+    internalMode = request.user.is_superuser
+    attributes = ['internal.area', 'internal.sub_area', 'cp_n_active_cc_users', 'deployment.countries']
+
+    if internalMode:
+        attributes = ['name', 'internal.organization_name', 'internal.notes'] + attributes
+
+    if 'country' in request.GET:
+        country = request.GET.get('country')
+        projects = (DomainES().is_active_project().real_domains()
+                    .filter(filters.term('deployment.countries', country))
+                    .sort('cp_n_active_cc_users', True).source(attributes).size(5).run().hits)
+        data = {country: projects, 'internal': internalMode}
+
+    return json_response(data)
+
+
+class DownloadMALTView(BaseAdminSectionView):
+    urlname = 'download_malt'
+    page_title = ugettext_lazy("Download MALT")
+    template_name = "hqadmin/malt_downloader.html"
+
+    @method_decorator(require_superuser)
+    def dispatch(self, request, *args, **kwargs):
+        return super(DownloadMALTView, self).dispatch(request, *args, **kwargs)
+
+    def get(self, request, *args, **kwargs):
+        from django.core.exceptions import ValidationError
+        if 'year_month' in request.GET:
+            try:
+                year, month = request.GET['year_month'].split('-')
+                year, month = int(year), int(month)
+                return _malt_csv_response(month, year)
+            except (ValueError, ValidationError):
+                messages.error(
+                    request,
+                    _("Enter a valid year-month. e.g. 2015-09 (for September 2015)")
+                )
+        return super(DownloadMALTView, self).get(request, *args, **kwargs)
+
+
+def _malt_csv_response(month, year):
+    query_month = "{year}-{month}-01".format(year=year, month=month)
+    queryset = MALTRow.objects.filter(month=query_month)
+    return export_as_csv_action(exclude=['id'])(MALTRowAdmin, None, queryset)
+
+
+class DownloadGIRView(BaseAdminSectionView):
+    urlname = 'download_gir'
+    page_title = ugettext_lazy("Download GIR")
+    template_name = "hqadmin/gir_downloader.html"
+
+    @method_decorator(require_superuser)
+    def dispatch(self, request, *args, **kwargs):
+        return super(DownloadGIRView, self).dispatch(request, *args, **kwargs)
+
+    def get(self, request, *args, **kwargs):
+        from django.core.exceptions import ValidationError
+        if 'year_month' in request.GET:
+            try:
+                year, month = request.GET['year_month'].split('-')
+                year, month = int(year), int(month)
+                return _gir_csv_response(month, year)
+            except (ValueError, ValidationError):
+                messages.error(
+                    request,
+                    _("Enter a valid year-month. e.g. 2015-09 (for September 2015)")
+                )
+        return super(DownloadGIRView, self).get(request, *args, **kwargs)
+
+
+def _gir_csv_response(month, year):
+    query_month = "{year}-{month}-01".format(year=year, month=month)
+    prev_month_year, prev_month = add_months(year, month, -1)
+    prev_month_string = "{year}-{month}-01".format(year=prev_month_year, month=prev_month)
+    two_ago_year, two_ago_month = add_months(year, month, -2)
+    two_ago_string = "{year}-{month}-01".format(year=two_ago_year, month=two_ago_month)
+    if not GIRRow.objects.filter(month=query_month).exists():
+        return HttpResponse('Sorry, that month is not yet available')
+    queryset = GIRRow.objects.filter(month__in=[query_month, prev_month_string, two_ago_string]).order_by('-month')
+    domain_months = defaultdict(list)
+    for item in queryset:
+        domain_months[item.domain_name].append(item)
+    field_names = GIR_FIELDS
+    response = HttpResponse(content_type='text/csv')
+    response['Content-Disposition'] = 'attachment; filename=gir.csv'
+    writer = UnicodeWriter(response)
+    writer.writerow(list(field_names))
+    for months in domain_months.values():
+        writer.writerow(months[0].export_row(months[1:]))
+    return response
